Create base class to share code
